`feat: update useUserCart hook and cart store to add product to cart`
